% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.CrI.width}
\alias{.CrI.width}
\title{CrI width}
\usage{
.CrI.width(CrI.low, CrI.high)
}
\arguments{
\item{CrI.low}{lower bound of credible interval}

\item{CrI.high}{upper bound of credible interval}
}
\value{
absolute distance between CrI and zero effect
}
\description{
CrI width
}
\keyword{internal}
