% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stan_draws.R
\name{extract_stan_draws}
\alias{extract_stan_draws}
\title{extract_stan_draws}
\usage{
extract_stan_draws(fit, parameter_name)
}
\arguments{
\item{fit}{A fitted Stan model object (e.g., \code{stanfit}, \code{brmsfit}, \code{stanreg}, \code{cmdstanfit}).}

\item{parameter_name}{A character string of parameter name}
}
\value{
A data frame with one row per MCMC draw and one column per parameter.
If multiple parameters, columns are named after the parameter.
}
\description{
Wrapper function to extract parameter draws from  from common Stan interfaces.
This function requires the respective stan interface (rstan, brms, cmdstanr, or rstanarm)
package to be installed.
If not available, you'll get an error.
}
\examples{
 #Not run:
 #fit <- brms::brm(count ~ zAge + zBase * Trt + (1|patient),

 # posterior <- extract_stan_draws(fit, "b_Intercept")
 # End(Not run)
}
