% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stan_fit.R
\name{extract_stan_fit}
\alias{extract_stan_fit}
\title{extract_stan_fit}
\usage{
extract_stan_fit(fit, parameter_name)
}
\arguments{
\item{fit}{A fitted Stan model object (\link[rstan]{stanfit},  \link[brms]{brmsfit}).}

\item{parameter_name}{A character string of parameter name}
}
\value{
A data frame with one row per MCMC draw and one column per parameter.
If multiple parameters, columns are named after the parameter.
}
\description{
Wrapper function to extract parameter draws from two common Stan interfaces.
This function requires the respective stan interface (rstan, brms)
package to be installed.
}
\examples{
 #Not run:
#fit <- brms::brm(count ~ zAge + zBase * Trt + (1|patient),
#             data = brms::epilepsy[1:30,], family = poisson())

 # posterior <- extract_stan_fit(fit, "b_Intercept")
 # End(Not run)
}
