% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plot Bayesian Volcano plot}
\usage{
plot_volcano(result, CrI = FALSE, CrI_width = FALSE, color = NULL)
}
\arguments{
\item{result}{from \code{\link[=prepare_volcano_input]{prepare_volcano_input()}} (a data frame).}

\item{CrI}{Logical. Whether to display the CrI Interval of the parameter}

\item{CrI_width}{Logical. Whether to display the CrI width as point size.}

\item{color}{Column in `result$result. Can be numerical or character.}
}
\value{
a ggplot2 object
}
\description{
Plot Bayesian Volcano plot
}
\examples{
data("posterior")
head(posterior)
data("annotation_df")
head(annotation_df)

result <- prepare_volcano_input(
  posterior = posterior,
  annotation = annotation_df,
)
plot_volcano(result,
  color = "group",
  CrI = TRUE,
  CrI_width = TRUE
)
}
