% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_volcano_input.R
\name{prepare_volcano_input}
\alias{prepare_volcano_input}
\title{Prepare volcano input}
\usage{
prepare_volcano_input(posterior, annotation, null.effect = 0, CrI_level = 0.95)
}
\arguments{
\item{posterior}{A data frame of posterior draws (one row per draw) \code{\link[=extract_fit]{extract_fit()}}.}

\item{annotation}{A data frame with at least one column:
\itemize{
\item \code{parameter}: the parameter name (e.g., \code{doubling.1}, \code{logOR.treatment})
\item \code{label}: the biological label (e.g., \code{cell.line}, \code{time.point})
\item Optional: other columns (e.g., \code{group}, \code{condition}) for future coloring
}}

\item{null.effect}{Central parameter value corresponding to no effect (default t=0).}

\item{CrI_level}{a scalar between 0 and 1 specifying the mass within the credible interval (default=0.95, i.e. 95\% credible interval (CrI)).}
}
\value{
A list with:
\itemize{
\item \code{result}: A data frame with columns:
\itemize{
\item \code{parameter}: original parameter name
\item \code{pi.value}: calculated pi.value
\item \code{null.effect}: set null effect by user
\item \code{parameter.median}: median posterior parameter value
\item \code{parameter.low}: lower boundary of CrI of parameter value
\item \code{paramter.high}: upper boundary of CrI of parameter value
\item \code{CrI.width}: the absolute distance between parameter.low and parameter.high
\item \code{CrI.level}: CrI_level set by user
\item \code{label}: biological label (e.g., \code{cell.line})
\item Other columns from \code{annotation} (e.g., \code{group}, \code{condition})
}
}
}
\description{
This function has as input posterior draws, calculates pi-values and credible intervals (CrI),
and annotates them with biological
information (e.g., cell line, time point) based on parameter names and a user-provided
annotation data frame. Returns a data frame that is ready for plotting.
}
\details{
Only returns pi-values and credible intervals for parameters that are \strong{in posterior and
annotation}. For formula see README or Vignette
}
\examples{
# Example: Simulate posterior and annotation
posterior <- data.frame(
  doubling.1 = rnorm(1000),
  doubling.2 = rnorm(1000)
)

annotation <- data.frame(
  parameter = c("doubling.1", "doubling.2"),
  label = c("cell.line.A", "cell.line.B"),
  group = c("group1", "group1")
)

result <- prepare_volcano_input(
  posterior = posterior,
  annotation = annotation,
)

head(result$result)

}
\seealso{
\code{\link[=extract_fit]{extract_fit()}}
}
