% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.pi_value}
\alias{.pi_value}
\title{pi_value}
\usage{
.pi_value(value, zero.effect)
}
\arguments{
\item{value}{numerical vector of posterior draws}

\item{zero.effect}{numerical value indicating for which treshhold to calculate pi}
}
\value{
pi-value
}
\description{
Calculates pi-value as integral under
}
\keyword{internal}
