% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Plot Bayesian Volcano plot}
\usage{
plot_volcano(
  result,
  CrI = FALSE,
  color = NULL,
  label = NULL,
  label.parameter.threshold = NULL,
  label.pi.threshold = NULL,
  title = "Bayesian Volcano Plot",
  xlab = "parameter value"
)
}
\arguments{
\item{result}{from \code{\link[=prepare_volcano_df]{prepare_volcano_df()}} (a list with \code{result} and \code{meta}).}

\item{CrI}{Logical. Whether to display the CrI Interval of the parameter}

\item{color}{Column in `result$result. Can be numerical or character.}

\item{label}{Character column name in \code{result$result} to use for labeling
points (e.g., "label", "parameter").
If \code{NULL}, no labels are shown.}

\item{label.parameter.threshold}{Numeric vector of length 1, absolute lower bound for label
If \code{NULL} or missing, all points are considered for labeling.}

\item{label.pi.threshold}{numeric vector of length 1, absolute lower bound for label
If \code{NULL} or missing, all points are considered for labeling.
for y-axis to trigger labeling.
If \code{NULL} or missing, all points are considered for labeling.}

\item{title}{Character. Title of plot}

\item{xlab}{Character. x-axis label of plot}
}
\value{
a ggplot2 object
}
\description{
Plot Bayesian Volcano plot
}
\examples{
data("posterior")
head(posterior)
data("annotation_df")
head(annotation_df)
 
result <- prepare_volcano_df(
  posterior = posterior,
  annotation_df = annotation_df,
)
plot_volcano(result,
            color="group",
            label.pi.threshold = 0.9,
            label.parameter.threshold = 0.5)
}
