% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_volcano_df.R
\name{prepare_volcano_df}
\alias{prepare_volcano_df}
\title{Prepare volcano data frame}
\usage{
prepare_volcano_df(
  posterior,
  annotation_df,
  zero.effect = 0,
  CrI.low = 0.025,
  CrI.high = 0.975
)
}
\arguments{
\item{posterior}{A data frame of posterior draws (one row per draw) \code{\link[=extract_fit]{extract_fit()}}.}

\item{annotation_df}{A data frame with at least one column:
\itemize{
\item \code{parameter}: the parameter name (e.g., \code{doubling.1}, \code{logOR.treatment})
\item \code{label}: the biological label (e.g., \code{cell.line}, \code{time.point})
\item Optional: other columns (e.g., \code{group}, \code{condition}) for future coloring
}}

\item{zero.effect}{Central parameter value corresponding to no effect (default t=0).}

\item{CrI.low}{lower bound of credible interval}

\item{CrI.high}{upper bound of credible interval}
}
\value{
A list with:
\itemize{
\item \code{result}: A data frame with columns:
\itemize{
\item \code{parameter}: original parameter name
\item \code{pi.value}: calulcated pi.value
\item \code{parameter.median}: median posterior parameter value
\item \code{parameter.low}: lower boundary of CrI of parameter value
\item \code{paramter.high}: upper boundary of CrI of parameter value
\item \code{CrI.width}: the absolute distance parameter.low and parameter.high
\item \code{label}: biological label (e.g., \code{cell.line})
\item Other columns from \code{annotation_df} (e.g., \code{group}, \code{condition})
}
\item \code{meta}: user settings
\itemize{
\item \code{CrI.low}: lower CrI boundary set by user
\item \code{CrI.high}: upper CrI boundary set by user
\item \code{zero.effect}: zero.effect for pi set by user
}
}
}
\description{
This function has as input posterior draws, calculates pi-values and credible intervals (CrI),
and annotates them with biological
information (e.g., cell line, time point) based on parameter names and a user-provided
annotation data frame. Returns a data frame that is ready for plotting.
}
\details{
Only returns pi-values and credible intervals for parameters that are \strong{in posterior and
annotation_df}. For formula see README or Vignette
}
\examples{
# Example: Simulate posterior and annotation
posterior <- data.frame(
  doubling.1 = rnorm(1000),
  doubling.2 = rnorm(1000)
)

annotation_df <- data.frame(
  parameter = c("doubling.1", "doubling.2"),
  label = c("cell.line.A", "cell.line.B"),
  group = c("group1", "group1")
)

result <- prepare_volcano_df(
  posterior = posterior,
  annotation_df = annotation_df,
)

head(result$result)

}
\seealso{
\code{\link[=extract_fit]{extract_fit()}}
}
